refactor: sepparate results component
