+++ react-components/src/components/results/results.css
@@ -8,27 +8,3 @@
 .searching {
   font-size: 100px;
 }
-
-.results__link {
-  border-radius: 5px;
-  border: 1px solid #b00020;
-  padding-left: 20px;
-  padding-right: 20px;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  margin-left: 10px;
-  margin-right: 10px;
-  text-decoration: none;
-  color: #b00020;
-}
-
-.results__link:hover {
-  transition: 0.3s;
-  transform: scale(1.2, 1.2);
-  box-shadow: 10px 5px 5px #282828;
-}
-
-.results__links__section {
-  margin-top: 20px;
-  display: flex;
-}
